Add a migration for moving the history data from task_entries to task_entries_history

+++ migrations/1642151908120_create-task-entries-history-table.js
@@ -28,6 +28,15 @@ exports.up = (pgm) => {
   })
 }
 
-exports.down = (pgm) => {
+exports.down = async (pgm) => {
+  // Make sure we don't have any data we can lose in the history table when doing a migrate down
+  const hasRows =
+    (await pgm.db.query('SELECT id FROM task_entries_history')).rows.length > 0
+  if (hasRows) {
+    throw new Error(`
+      Migrate down aborted - possible lose of data if task_entries_history table is dropped: \n 
+      Make sure you've moved all the history data from task_entries_history to task_entries and emptied the task_entries_history table.
+    `)
+  }
   pgm.dropTable('task_entries_history')
 }
